backend/requirements.txt
# Flask Web Framework & Ecosystem
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.0
Flask-Marshmallow>=0.14
marshmallow-sqlalchemy>=0.24
python-dotenv>=0.19
Werkzeug>=2.0 # For file uploads securely
gunicorn # Recommended for production deployment (optional for dev)

# Database Driver (PostgreSQL)
psycopg2-binary # Or psycopg2 if build tools are available

# Celery for Background Tasks
celery[redis]>=5.0 # Includes redis dependency
redis>=3.5 # Redis client library

# NLP & Resume Processing
spacy>=3.0        # Core NLP library
scikit-learn>=1.0 # For TF-IDF, Cosine Similarity
nltk>=3.6         # For stopwords, potentially other NLP tasks
PyPDF2>=3.0       # PDF parsing
python-docx>=0.8  # DOCX parsing
textract>=1.6     # Optional fallback parser (needs OS dependencies)

# Semantic Similarity (Requires Transformers & PyTorch/TensorFlow)
sentence-transformers>=2.2.0
transformers>=4.20.0 # Dependency for sentence-transformers
# torch>=1.9 # Or tensorflow>=2.5 - Sentence-Transformers usually pulls one in
tf-keras # Compatibility layer for Keras 3 if using TensorFlow backend

# Optional: Add linters/formatters if desired for development
# black
# flake8